Update unit test 04, delete inference
