--- inference/README.md
@@ -1,3 +1,0 @@
-# Inférence
-
-Fonction de prédiction utilisée dans les phases de tests.
